requirements.txt
#with the requirements file you can define exactly which versions you were using
#this helps in trouble shooting cases

ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
# kafka==1.3.5 # not sure if this library is also required
kafka-python==2.0.2
numpy==1.23.1
pandas==1.4.3
plotly==5.9.0
requests==2.25.1

# Python version: 3.8.6